LPS-126044 Move code to a better module.
