+++ modules/dxp/apps/analytics/analytics-reports-api/src/main/resources/com/liferay/analytics/reports/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
