+++ modules/dxp/apps/analytics/analytics-reports-layout-display-page-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Analytics Reports Layout Display Page Implementation
+Bundle-SymbolicName: com.liferay.analytics.reports.layout.display.page.impl
+Bundle-Version: 1.0.0
