+++ modules/dxp/apps/analytics/analytics-reports-layout-display-page-impl/build.gradle
@@ -1,0 +1,11 @@
+dependencies {
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
+	compileOnly group: "org.osgi", name: "org.osgi.annotation.versioning", version: "1.1.0"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
+	compileOnly project(":apps:info:info-api")
+	compileOnly project(":apps:layout:layout-display-page-api")
+	compileOnly project(":core:petra:petra-string")
+	compileOnly project(":dxp:apps:analytics:analytics-reports-api")
+}
