+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/init.jsp
@@ -25,7 +25,7 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %><%@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
-<%@ page import="com.liferay.analytics.reports.web.internal.constants.AnalyticsReportsWebKeys" %><%@
+<%@ page import="com.liferay.analytics.reports.constants.AnalyticsReportsWebKeys" %><%@
 page import="com.liferay.analytics.reports.web.internal.display.context.AnalyticsReportsDisplayContext" %><%@
 page import="com.liferay.analytics.reports.web.internal.util.AnalyticsReportsUtil" %><%@
 page import="com.liferay.portal.kernel.util.PortalUtil" %><%@
